Se hace correcion de las pruebas de responses
