+++ tests/resources/responses/response_getstatusZip_with_errors.xml
@@ -81,7 +81,7 @@
                     <b:XmlBytes i:nil="true"/>
                     <b:XmlDocumentKey>840ca7a4606904acb5d979db78257876f7c987261cb499e9b4a4c4fd18864db9a7c54c1b15affb68997550889cffd4f7
                     </b:XmlDocumentKey>
-                    <b:XmlFileName>face_f08300448580000000003</b:XmlFileName>
+                    <b:XmlFileName>face_f07300448580000000003</b:XmlFileName>
                 </b:DianResponse>
             </GetStatusZipResult>
         </GetStatusZipResponse>
